added final language fixes
