+++ env.example
@@ -8,4 +8,6 @@ NOT_STARTED_TRANSLATIONS=hu,ru
 IN_REVIEW_TRANSLATIONS=ja
 
 DEFAULT_LOCALE=en
+LOCALES=ja,hu,ru,en
+
 ENV=development
